Moved tests into tests package
