+++ PluginProject/src/test/HalsteadLengthCheckTest.java
@@ -1,8 +1,10 @@
-package StructuralMetrics;
+package test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+
+import StructuralMetrics.HalsteadLengthCheck;
 
 class HalsteadLengthCheckTest {
